Implemented Friendship Events to Kafka

+++ src/main/java/com/pingr/Connections/core/Friendships/FriendshipService.java
@@ -1,5 +1,6 @@
 package com.pingr.Connections.core.Friendships;
 
+import com.pingr.Connections.application.ProducerService;
 import com.pingr.Connections.core.Accounts.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,18 +11,15 @@ import java.util.Optional;
 
 @Service
 public class FriendshipService {
-    private FriendshipRepository friendshipRepo;
-    private AccountRepository accountRepo;
+    private final FriendshipRepository friendshipRepo;
+    private final AccountRepository accountRepo;
+    private final ProducerService producer;
 
     @Autowired
-    public FriendshipService(FriendshipRepository friendshipRepo, AccountRepository accountRepo) {
+    public FriendshipService(FriendshipRepository friendshipRepo, AccountRepository accountRepo, ProducerService producer) {
         this.friendshipRepo = friendshipRepo;
         this.accountRepo = accountRepo;
-    }
-
-    public boolean areAlreadyFriends (Friendship friendship) {
-        Optional<Friendship> opt = friendshipRepo.findFriendship(friendship.getIdAccountApplied(), friendship.getIdAccountReceived());
-        return opt.isPresent();
+        this.producer = producer;
     }
 
     public List<Friendship> findFriendsOfOneAccount(Long idAccount) {
@@ -39,14 +37,29 @@ public class FriendshipService {
         if (areAlreadyFriends(friendship))
             throw new IllegalArgumentException("They are already friends");
 
-        boolean accountsExist = accountRepo.existsById(friendship.getIdAccountApplied()) && accountRepo.existsById(friendship.getIdAccountReceived());
-        if(!accountsExist)
+        if(! accountsExist(friendship))
             throw new IllegalArgumentException("Account id not found");
 
-        return friendshipRepo.save(friendship);
+        Friendship savedFriendship = friendshipRepo.save(friendship);
+        this.producer.emitFriendshipCreatedEvent(savedFriendship);
+        return savedFriendship;
     }
 
     public void cancel(Long idAccount1, Long idAccount2) {
+        Optional<Friendship> optFriendship = friendshipRepo.findFriendship(idAccount1, idAccount2);
+        if (optFriendship.isEmpty())
+            throw new IllegalArgumentException("There is no friendship between these IDs");
+
+        this.producer.emitFriendshipDeletedEvent(optFriendship.get());
         friendshipRepo.deleteFriendship(idAccount1, idAccount2);
+    }
+
+    private boolean areAlreadyFriends (Friendship friendship) {
+        Optional<Friendship> opt = friendshipRepo.findFriendship(friendship.getIdAccountApplied(), friendship.getIdAccountReceived());
+        return opt.isPresent();
+    }
+
+    private boolean accountsExist(Friendship friendship) {
+        return accountRepo.existsById(friendship.getIdAccountApplied()) && accountRepo.existsById(friendship.getIdAccountReceived());
     }
 }
